requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-filter==24.3
python-dotenv==1.0.1
PyMySQL==1.1.1
